requirements.txt
# Internet usage tracker — Ethernet only
psutil>=7.2.2
scapy>=2.7.0
pystray>=0.19.0
Pillow>=10.0.0
